requirements.txt
streamlit==1.44.1
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-chroma==0.1.4
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
tenacity==8.5.0
pandas==2.2.3
ragas==0.1.21
datasets==2.21.0
